pyproject.toml
[project]
name = "rush"
version = "0.1.0"
description = "Swiss public transport and weather analytics pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.7",
    "dbt-bigquery>=1.7",
    "dbt-postgres>=1.7",
    "dlt[postgres]>=1.0",
    "google-cloud-storage>=2.14",
    "pandas>=2.2",
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "requests>=2.31",
    "sqlalchemy>=2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2",
    "jupyter>=1.1",
    "pgcli>=4.4",
    "ruff>=0.9",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "double"
